remove setup and move request to separate module
