+++ python/.flake8
@@ -1,0 +1,16 @@
+
+[flake8]per-file-ignores =
+    __init__.py:
+        # import but unused
+        F401
+        # unable to detect undefined names
+        F403
+exclude =
+    venv,
+    .git,
+    __pycache__,
+    .vscode,
+    .idea,
+    dashy.egg-info
+max-complexity = 10
+max-line-length = 127
